Fix: Class "CompanySubscription" has 24 methods, which is greater than 20 authorized. Split it into smaller classes.
