+++ app/Models/CompanySubscription.php
@@ -2,6 +2,10 @@
 
 namespace App\Models;
 
+use App\Models\Concerns\HasSubscriptionActions;
+use App\Models\Concerns\HasSubscriptionDisplay;
+use App\Models\Concerns\HasSubscriptionStatus;
+use App\Models\Concerns\ManagesUserLimits;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -35,6 +39,10 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
 class CompanySubscription extends Model
 {
     use HasFactory;
+    use HasSubscriptionActions;
+    use HasSubscriptionDisplay;
+    use HasSubscriptionStatus;
+    use ManagesUserLimits;
 
     /**
      * The table associated with the model.
@@ -114,154 +122,6 @@ class CompanySubscription extends Model
     }
 
     /**
-     * Check if the subscription is active.
-     */
-    public function isActive(): bool
-    {
-        return in_array($this->status, [self::STATUS_ACTIVE, self::STATUS_TRIALING]);
-    }
-
-    /**
-     * Check if the subscription is on trial.
-     */
-    public function onTrial(): bool
-    {
-        return $this->status === self::STATUS_TRIALING &&
-               $this->trial_ends_at &&
-               $this->trial_ends_at->isFuture();
-    }
-
-    /**
-     * Check if the subscription is canceled.
-     */
-    public function isCanceled(): bool
-    {
-        return $this->status === self::STATUS_CANCELED;
-    }
-
-    /**
-     * Check if the subscription is suspended.
-     */
-    public function isSuspended(): bool
-    {
-        return $this->status === self::STATUS_SUSPENDED;
-    }
-
-    /**
-     * Check if the subscription is within grace period.
-     */
-    public function onGracePeriod(): bool
-    {
-        return $this->grace_period_ends_at &&
-               $this->grace_period_ends_at->isFuture();
-    }
-
-    /**
-     * Check if the company can add more users.
-     * This excludes client portal users from the count.
-     */
-    public function canAddUser(): bool
-    {
-        // If no plan or unlimited users
-        if (! $this->subscriptionPlan || $this->max_users === null) {
-            return true;
-        }
-
-        // Check if under limit
-        return $this->current_user_count < $this->max_users;
-    }
-
-    /**
-     * Get the number of available user slots.
-     */
-    public function availableUserSlots(): ?int
-    {
-        if ($this->max_users === null) {
-            return null; // Unlimited
-        }
-
-        return max(0, $this->max_users - $this->current_user_count);
-    }
-
-    /**
-     * Update the current user count.
-     * This should only count non-portal users.
-     */
-    public function updateUserCount(): void
-    {
-        $count = User::where('company_id', $this->company_id)
-            ->whereNull('archived_at')
-            ->where('status', true)
-            ->count();
-
-        // Use updateQuietly to prevent triggering model events and avoid infinite loop
-        $this->updateQuietly(['current_user_count' => $count]);
-    }
-
-    /**
-     * Check if a feature is available in the subscription.
-     */
-    public function hasFeature(string $feature): bool
-    {
-        if (! $this->features) {
-            return false;
-        }
-
-        return in_array($feature, $this->features) ||
-               (isset($this->features[$feature]) && $this->features[$feature] === true);
-    }
-
-    /**
-     * Get the subscription display name.
-     */
-    public function getDisplayName(): string
-    {
-        if ($this->subscriptionPlan) {
-            return $this->subscriptionPlan->name;
-        }
-
-        return 'No Plan';
-    }
-
-    /**
-     * Get the subscription price display.
-     */
-    public function getPriceDisplay(): string
-    {
-        if ($this->monthly_amount == 0) {
-            return 'Free';
-        }
-
-        return '$'.number_format($this->monthly_amount, 2).'/month';
-    }
-
-    /**
-     * Get user limit display text.
-     */
-    public function getUserLimitDisplay(): string
-    {
-        if ($this->max_users === null) {
-            return 'Unlimited users';
-        }
-
-        return $this->current_user_count.' of '.$this->max_users.' users';
-    }
-
-    /**
-     * Check if approaching user limit (80% or more).
-     */
-    public function approachingUserLimit(): bool
-    {
-        if ($this->max_users === null) {
-            return false;
-        }
-
-        $percentage = ($this->current_user_count / $this->max_users) * 100;
-
-        return $percentage >= 80;
-    }
-
-    /**
      * Scope to get active subscriptions.
      */
     public function scopeActive($query)
@@ -285,94 +145,6 @@ class CompanySubscription extends Model
     {
         return $query->where('status', self::STATUS_TRIALING)
             ->whereBetween('trial_ends_at', [now(), now()->addDays($days)]);
-    }
-
-    /**
-     * Cancel the subscription.
-     */
-    public function cancel(bool $immediately = false): void
-    {
-        if ($immediately) {
-            $this->update([
-                'status' => self::STATUS_CANCELED,
-                'canceled_at' => now(),
-            ]);
-        } else {
-            // Cancel at end of period
-            $this->update([
-                'canceled_at' => now(),
-                'grace_period_ends_at' => $this->current_period_end,
-            ]);
-        }
-    }
-
-    /**
-     * Suspend the subscription.
-     */
-    public function suspend(?string $reason = null): void
-    {
-        $metadata = $this->metadata ?? [];
-        if ($reason) {
-            $metadata['suspension_reason'] = $reason;
-        }
-
-        $this->update([
-            'status' => self::STATUS_SUSPENDED,
-            'suspended_at' => now(),
-            'metadata' => $metadata,
-        ]);
-    }
-
-    /**
-     * Resume a suspended subscription.
-     */
-    public function resume(): void
-    {
-        $metadata = $this->metadata ?? [];
-        unset($metadata['suspension_reason']);
-
-        $this->update([
-            'status' => self::STATUS_ACTIVE,
-            'suspended_at' => null,
-            'metadata' => $metadata,
-        ]);
-    }
-
-    /**
-     * Change the subscription plan.
-     */
-    public function changePlan(SubscriptionPlan $newPlan): void
-    {
-        $this->update([
-            'subscription_plan_id' => $newPlan->id,
-            'max_users' => $newPlan->max_users,
-            'monthly_amount' => $newPlan->price_monthly,
-            'features' => $newPlan->features,
-        ]);
-
-        // If downgrading and over new limit, mark for attention
-        if ($newPlan->max_users && $this->current_user_count > $newPlan->max_users) {
-            $metadata = $this->metadata ?? [];
-            $metadata['over_user_limit'] = true;
-            $metadata['previous_user_count'] = $this->current_user_count;
-            $this->update(['metadata' => $metadata]);
-        }
-    }
-
-    /**
-     * Get status badge color for UI.
-     */
-    public function getStatusColor(): string
-    {
-        return match ($this->status) {
-            self::STATUS_ACTIVE => 'green',
-            self::STATUS_TRIALING => 'blue',
-            self::STATUS_PAST_DUE => 'orange',
-            self::STATUS_CANCELED => 'gray',
-            self::STATUS_SUSPENDED => 'red',
-            self::STATUS_EXPIRED => 'gray',
-            default => 'gray',
-        };
     }
 
     /**
